factor out parsing of json to function 'parse'
parse outputs both the file info and the transaction info.

the 'info' and 'csv' functions
discard the parts they don't need

future functions calling parse will use both parts,
specifically, a function that
generates the transaction info,
then writes it to a file
with a name based on the file info.