Substantial rewrite
I've stopped trying to stream everything,
which I think is pointless
because I think jq holds everything in memory.
So now I'm putting things in variables.
One for the xml-ified ofx,
one for the resulting json,
and another for the tweaked json
with things in easy-to-reach places
and with more derived properties.
this last one is then used as the source
when generating output information, transactions, and filenames.